pyproject.toml
[project]
name = "django-sns-view"
version = "0.1.2"
description = "A Django view that can be subscribed to Amazon SNS"
authors = [
    {name = "thelab", email = "thelabdev@thelab.co"}
]
license = {text = "ISC"}
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "Django (>=4.2)",
    "pyOpenSSL (>=24.2.1,<24.3.0)",  # 24.3.0 removes openssl.verify
    "pem (>=23.1.0)",
    "requests (>=2.32.3)",
    "pydantic (>=2.10.6,<3.0.0)",
]

[tool.poetry]
packages = [
    { include = "django_sns_view" },
]

[tool.poetry.group.dev.dependencies]
mypy = "^1.15.0"
ruff = "^0.9.10"
coverage = "^7.6.12"
django-stubs = "^5.1.3"
types-cryptography = "^3.3.23.2"
types-requests = "^2.32.0.20250306"
tox = "^4.24.2"
monkeytype = "^23.3.0"
types-pyOpenSSL = "^24.1.0.20240722"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
from_first = true

[tool.ruff.lint.isort]
from-first = true

[tool.mypy]
python_version = "3.12"
plugins = ["mypy_django_plugin.main"]

# Strict mode, see mypy --help
warn_unused_configs = true
disallow_subclassing_any = true
disallow_any_generics = true
disallow_untyped_calls = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_return_any = true
no_implicit_reexport = true
show_error_codes = true
# Not turned on by strict
strict_equality = true

[tool.django-stubs]
django_settings_module = "django_sns_view.tests.settings"

[tool.coverage.run]
branch = true
source_pkgs = ["django_sns_view"]
omit = [
    "*/django_sns_view/tests/*",
]

[tool.coverage.report]
show_missing = true
ignore_errors = true

[tool.commitizen]
name = "cz_conventional_commits"
annotated_tag = true
gpg_sign = true
tag_format = "v$version"
update_changelog_on_bump = true
changelog_merge_prerelease = true
version_provider = "pep621"
version_scheme = "pep440"
version_files = [
    "pyproject.toml:version",
]
pre_bump_hooks = [
    "pre-commit run --all-files || true",
]
post_bump_hooks = [
    "git push origin master $CZ_POST_CURRENT_TAG_VERSION"
]
